ENH: License the `mbs.css` API under terms of LGPL3
