+++ src/api/css.ts
@@ -1,4 +1,19 @@
-import { fromEntries, entries, logger } from "../common";
+/**
+ * Copyright (C) Bananarama92
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of
+ * the GNU Lesser General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along with this program.
+ * If not, see https://www.gnu.org/licenses/.
+ */
+
+import { fromEntries, entries, logger } from "../lpgl/common";
 
 /** A record mapping {@link DEFAULT_STYLE} keys to their CSS equivalent (see the `index.scss` root selector) */
 const FIELD_MAPPING = Object.freeze({
