+++ src/lpgl/README.md
@@ -1,17 +1,12 @@
-MBS: Maid's Bondage Scripts
-
 Copyright (C) 2023-2024 Bananarama92
 
 This program is free software: you can redistribute it and/or modify it under the terms of
-the GNU General Public License as published by the Free Software Foundation,
+the GNU Lesser General Public License as published by the Free Software Foundation,
 either version 3 of the License, or (at your option) any later version.
 
 This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
 without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-See the GNU General Public License for more details.
+See the GNU Lesser General Public License for more details.
 
 You should have received a copy of the GNU General Public License along with this program.
 If not, see https://www.gnu.org/licenses/.
-
-An exception is made for the `mbs.css` API, which is licensed under the terms of GNU Lesser
-General Public License version 3 (see https://www.gnu.org/licenses/lgpl-3.0.en.html).
